requirements.txt
ops==3.3.0
charmlibs.snap==1.0.0
